テクスチャ差し替え、ブロックの耐久値をfloatに変更、ブロックの耐久力ごとに色分けをした

+++ DirectXGame/Game/Player/Player.h
@@ -269,7 +269,7 @@ public:
 	}
 
 	//パワーーー！！！を強化
-	void UpgradePower(int32_t addNum) { parameters_[currentCharacters_]->dig_.digPower += addNum; }
+	void UpgradePower(float addNum) { parameters_[currentCharacters_]->dig_.digPower += addNum; }
 
 	//ブロックの数取得
 	int32_t GetRockCount() const { return rockParameter_.rocks_[BringRocks::kRock]; }
